+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.9.0] - 2021-05-26
 ### Added
 - Allow `pq()` helper function to print bare numbers (i.e. without units).
 - Add `pf()` and `pi()` helpers to print floats and integers.
@@ -56,6 +58,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial public release
 
 [Unreleased]: https://github.com/gbingersoll/sciengdox/compare/v0.9.0...HEAD
+[0.9.0]: https://github.com/gbingersoll/sciengdox/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/gbingersoll/sciengdox/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/gbingersoll/sciengdox/compare/v0.6.3...v0.7.0
 [0.6.3]: https://github.com/gbingersoll/sciengdox/compare/v0.6.2...v0.6.3